Add weekly work
No commits during the week because of major restructure stuff. Needed
to restructure previous transformations to work with the new pipeline
system.

The transpiler now operates according to pipelines, which are made up of
a sequence of stages. Each stage can define a main function that gets
the output of the previous stage as parameters at call-time, and its
output is subsequently forwarded to the next stage in a similar manner.

The GLSL pipeline right now includes:
- preprocessor:
    this performs AST -> AST transformations that make later processing
    easier
- scope analysis
    this determines the bounds of each scope in the Julia scope and
    marks each AST node accordingly
- symbol resolution
    this stage maps each symbol to a unique identifier, that can later
    be used in the generated GLSL code, without having to worry about
    GLSL scoping rules
- type inference
    this stage marks each AST node with its type
    it has not been reworked to fit the new code structure, so it's
    unavailable for now
    although most of the logic for this is present in the current code,
    it will have to be ported over to use the new pipeline system