Add interface variables, VecNT support and swizzle
In, out and uniform variables can now be specified using the @in, @out
and @uniform macros respectively in the transpiled function's parameter
list.
To facilitate this, pipelines now have pipeline contexts, which are
automatically passed to and between stages. Each stage can choose to
use it or ignore it. This way, pipeline-level data can be kept without
having to adjust the generic target language agnostic stages in the
middle.

The VecNT types from the glm.jl lib are now also supported, for every
element type and count. For now, the transpilation environment must use
the types that are pulled in from the glm lib found in the transpiler's
module, but I'm looking into options to eliminate this restriction.

Broadcasted calls (e.g. cos.(v::Vec3)) are also interpreted properly,
and will be present simply as cos(v) in the generated code. There's no
logic for now to check whether the generated function call is actually
valid, so as long as it's a valid broadcasted fn call in the Julia
environment, the transpiler will assume it has a fn in glsl with the
same name.