Add env syms and more broadcasting support
Pipelines can now define their own quasi-locals through env syms. These
are meant for variables that are meant to be parsed mostly as-is and
have a pre-determined type.
In GLSL, these are mostly the gl_ variables, although for now the list
only contains the minimal stuff needed. In/out/uniform variables also
work through this now.