Add new approach to built-in functions
Added a bunch of GLSL built-in functions with type signatures that can
easily be expanded later. The signatures are stored in a Dict and can be
defined with some helper functions. Those signatures are than evaluated
to be method overloads to a central builtin_fn_ret_type function in the
TypeInference module, which is then used by the transpiler.

Also some bug fixes and WIP stuff on things like forced fn call
rewriting for JuliaGLM function calls.