Refactor, comment and write README.md
Refactored some of the code and folder structure to be more concise and
more intuitive.

Added comments to most of the core functions, although the core pipeline
could probably benefit from some more docstrings and comments.
Most of that code should be self-documenting mostly though, because
every tree transformation function carries the information about what it
handles exactly in its signature.

Also wrote a README.md for understanding the project and its structure
for further development.